+++ config/application.rb
@@ -5,7 +5,7 @@ require "rails/all"
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
-Dotenv::Railtie.load unless ["production", "staging"].include?(ENV["RAILS_ENV"])
+Dotenv::Railtie.load unless ['production', 'staging'].include?(ENV['RAILS_ENV'])
 
 module CtdSales
   class Application < Rails::Application