+++ config/initializers/shopify.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-ShopifyAPI::Base.site = ENV.fetch("SHOPIFY_URL", Rails.application.credentials[:shopify][:url])
+ShopifyAPI::Base.site = ENV.fetch("SHOPIFY_URL")
 ShopifyAPI::Base.api_version = "2020-01"